Splits the consumers and redis calls to different files (WIP)
