+++ .gitignore
@@ -6,3 +6,4 @@ database/
 secrets/
 backend/mediafiles/profile_pics/
 .venv/
+Pong.lwp
