+++ backend/pong/routing.py
@@ -1,5 +1,5 @@
 from django.urls import re_path
-from .consumers import MatchmakingConsumer, GameSessionConsumer
+from .consumers.consumers import MatchmakingConsumer, GameSessionConsumer
 
 websocket_urlpatterns = [
     re_path(r"ws/pong/matchmaking/$", MatchmakingConsumer.as_asgi()),
